feat: Refactor connection credentials page
