+++ src/Resources/app/administration/src/module/swag-migration/index.js
@@ -119,13 +119,6 @@ Module.register('swag-migration', {
                         privilege: 'admin',
                     },
                 },
-                profileInformation: {
-                    path: 'profile/information',
-                    component: 'swag-migration-wizard-page-profile-information',
-                    meta: {
-                        privilege: 'admin',
-                    },
-                },
                 credentials: {
                     path: 'credentials',
                     component: 'swag-migration-wizard-page-credentials',
