+++ src/Resources/app/administration/src/module/swag-migration/page/wizard/index.js
@@ -8,7 +8,6 @@ import './swag-migration-wizard-page-introduction';
 import './swag-migration-wizard-page-profile-installation';
 import './swag-migration-wizard-page-connection-select';
 import './swag-migration-wizard-page-connection-create';
-import './swag-migration-wizard-page-profile-information';
 import './swag-migration-wizard-page-credentials';
 import './swag-migration-wizard-page-credentials-success';
 import './swag-migration-wizard-page-credentials-error';
