--- src/Resources/app/administration/src/module/swag-migration/page/wizard/swag-migration-wizard-page-profile-information/index.js
@@ -1,30 +1,0 @@
-import template from './swag-migration-wizard-page-profile-information.html.twig';
-
-const { Component } = Shopware;
-
-/**
- * @private
- * @sw-package fundamentals@after-sales
- */
-Component.register('swag-migration-wizard-page-profile-information', {
-    template,
-
-    props: {
-        profileInformationComponent: {
-            type: String,
-            default: '',
-        },
-    },
-
-    metaInfo() {
-        return {
-            title: this.$createTitle(),
-        };
-    },
-
-    computed: {
-        componentIsLoaded() {
-            return Component.getComponentRegistry().has(this.profileInformationComponent);
-        },
-    },
-});
