--- src/Resources/app/administration/src/module/swag-migration/page/wizard/swag-migration-wizard-page-profile-information/swag-migration-wizard-page-profile-information.html.twig
@@ -1,9 +1,0 @@
-{% block swag_migration_wizard_page_profile_information %}
-    <div>
-
-        {% block swag_migration_wizard_page_profile_information_component %}
-            <component :is="profileInformationComponent"
-                       v-if="profileInformationComponent !== '' && componentIsLoaded" />
-        {% endblock %}
-    </div>
-{% endblock %}
