+++ src/Resources/app/administration/src/module/swag-migration/profile/shopware/api/swag-migration-profile-shopware-api-credential-form/index.js
@@ -56,6 +56,24 @@ Component.register('swag-migration-profile-shopware-api-credential-form', {
                 },
             });
         },
+
+        storeLink() {
+            return `https://store.shopware.com/${this.storeLinkISO}/swag226607479310f/migration-connector.html`;
+        },
+
+        storeLinkISO() {
+            const iso = this.locale.split('-')[0];
+
+            if (['en', 'de'].includes(iso)) {
+                return iso;
+            }
+
+            return 'en';
+        },
+
+        locale() {
+            return Shopware.Store.get('session').currentLocale ?? '';
+        },
     },
 
     watch: {
