+++ src/Resources/app/administration/src/module/swag-migration/profile/shopware/api/swag-migration-profile-shopware-api-credential-form/swag-migration-profile-shopware-api-credential-form.html.twig
@@ -1,16 +1,27 @@
 <div class="swag-migration-wizard swag-migration-wizard-page-credentials">
 
     {% block swag_migration_wizard_page_credentials_content %}
-        <div class="swag-migration-wizard__content">
 
             {% block swag_migration_wizard_page_credentials_information %}
                 <div class="swag-migration-wizard__content-information">
-
-                    {% block swag_migration_wizard_page_credentials_api_hint %}
-                        <mt-banner variant="attention">
-                            <p v-html="$t('swag-migration.wizard.pages.credentials.shopware55.api.contentInformation')"></p>
-                        </mt-banner>
-                    {% endblock %}
+                    <h3 class="text-left">
+                        {{ $tc('swag-migration.wizard.pages.credentials.shopware55.api.title') }}
+                    </h3>
+                    <ol class="swag-migration-wizard-page-credentials__steps">
+                        <li>
+                            <span v-html="$tc('swag-migration.wizard.pages.credentials.shopware55.api.steps.step1')"></span>
+                            <a class="link"
+                               target="_blank"
+                               :href="storeLink">
+                                {{ $tc('swag-migration.wizard.pages.credentials.shopware55.api.pluginLinkText') }}
+                            </a>
+                        </li>
+                        <li>{{ $tc('swag-migration.wizard.pages.credentials.shopware55.api.steps.step2') }}</li>
+                        <li>{{ $tc('swag-migration.wizard.pages.credentials.shopware55.api.steps.step3') }}</li>
+                        <li>{{ $tc('swag-migration.wizard.pages.credentials.shopware55.api.steps.step4') }}</li>
+                        <li>{{ $tc('swag-migration.wizard.pages.credentials.shopware55.api.steps.step5') }}</li>
+                        <li>{{ $tc('swag-migration.wizard.pages.credentials.shopware55.api.steps.step6') }}</li>
+                    </ol>
                 </div>
             {% endblock %}
 
@@ -52,15 +63,15 @@
                 {% endblock %}
 
                 {% block swag_migration_wizard_page_credentials_endpoint_field %}
-                    <mt-text-field
+                    <mt-url-field
                                   v-model:model-value="inputCredentials.endpoint"
                                   :validation="inputCredentials.endpoint !== 'http://' && inputCredentials.endpoint !== 'https://'"
                                   :label="$tc('swag-migration.wizard.pages.credentials.shopware55.api.endpointLabel')"
                                   :placeholder="$tc('swag-migration.wizard.pages.credentials.shopware55.api.endpointPlaceholder')"
-                                  :switch-label="$tc('swag-migration.wizard.pages.credentials.shopware55.api.endpointSSLLabel')" />
+                                  :switch-label="$tc('swag-migration.wizard.pages.credentials.shopware55.api.endpointSSLLabel')"
+                                  required />
                 {% endblock %}
             </div>
             {% endblock %}
-        </div>
     {% endblock %}
 </div>
