+++ src/Resources/app/administration/src/module/swag-migration/profile/shopware6/api/swag-migration-profile-shopware6-api-credential-form/swag-migration-profile-shopware6-api-credential-form.html.twig
@@ -1,16 +1,32 @@
 <div class="swag-migration-wizard swag-migration-wizard-page-credentials">
 
     {% block swag_migration_wizard_page_credentials_content %}
-        <div class="swag-migration-wizard__content">
 
             {% block swag_migration_wizard_page_credentials_information %}
                 <div class="swag-migration-wizard__content-information">
-
-                    {% block swag_migration_wizard_page_credentials_api_hint %}
-                        <p class="swag-migration-wizard__description">
-                            {{ $tc('swag-migration.wizard.pages.credentials.shopware6.api.contentInformation') }}
-                        </p>
-                    {% endblock %}
+                    <h3 class="text-left">
+                        {{ $tc('swag-migration.wizard.pages.credentials.shopware6.api.title') }}
+                    </h3>
+                    <ol class="swag-migration-wizard-page-credentials__steps">
+                        <li>
+                            <span v-html="$tc('swag-migration.wizard.pages.credentials.shopware6.api.steps.step1')"></span>
+                            <a class="link"
+                               target="_blank"
+                               :href="$tc('swag-migration.wizard.pages.credentials.shopware6.api.pluginLink')">
+                                {{ $tc('swag-migration.wizard.pages.credentials.shopware6.api.pluginLinkText') }}
+                            </a>
+                        </li>
+                        <li>
+                            {{ $tc('swag-migration.wizard.pages.credentials.shopware6.api.steps.step2') }}
+                            <a class="link"
+                               target="_blank"
+                               :href="$tc('swag-migration.wizard.pages.credentials.shopware6.api.integrationLink')">
+                                {{ $tc('swag-migration.wizard.pages.credentials.shopware6.api.integrationLinkText') }}
+                            </a>
+                        </li>
+                        <li>{{ $tc('swag-migration.wizard.pages.credentials.shopware6.api.steps.step3') }}</li>
+                        <li>{{ $tc('swag-migration.wizard.pages.credentials.shopware6.api.steps.step4') }}</li>
+                    </ol>
                 </div>
             {% endblock %}
 
@@ -57,6 +73,5 @@
                     {% endblock %}
                 </div>
             {% endblock %}
-        </div>
     {% endblock %}
 </div>
