--- src/Resources/app/administration/src/module/swag-migration/profile/shopware6/api/swag-migration-profile-shopware6-api-page-information/index.js
@@ -1,12 +1,0 @@
-import template from './swag-migration-profile-shopware6-api-page-information.html.twig';
-import './swag-migration-profile-shopware6-api-page-information.scss';
-
-const { Component } = Shopware;
-
-/**
- * @private
- * @sw-package fundamentals@after-sales
- */
-Component.register('swag-migration-profile-shopware6major-api-page-information', {
-    template,
-});
