--- src/Resources/app/administration/src/module/swag-migration/profile/shopware6/api/swag-migration-profile-shopware6-api-page-information/swag-migration-profile-shopware6-api-page-information.html.twig
@@ -1,65 +1,0 @@
-{% block swag_migration_wizard_page_plugin_information_shopware6 %}
-    <div class="swag-migration-wizard swag-migration-wizard-page-plugin-information swag-migration-profile-shopware6-api-page-information">
-
-        {% block swag_migration_wizard_page_plugin_information_shopware6_content %}
-            <div class="swag-migration-wizard__content">
-
-                {% block swag_migration_wizard_page_plugin_information_shopware6_information %}
-                    <div class="swag-migration-wizard__content-information">
-
-                        {% block swag_migration_wizard_page_plugin_information_shopware6_icon %}
-                            <div class="swag-migration-profile-shopware6-api-page-information__icon sw-settings-item">
-                                <div class="sw-settings-item__icon background--enabled">
-                                    <mt-icon name="regular-cog" />
-                                </div>
-                            </div>
-                        {% endblock %}
-
-                        {% block swag_migration_wizard_page_plugin_information_shopware6_title %}
-                            <h3 class="information__header">
-                                {{ $tc('swag-migration.wizard.pages.profileInformation.shopware6Api.contentTitle') }}
-                            </h3>
-                        {% endblock %}
-
-                        {% block swag_migration_wizard_page_plugin_information_shopware6_api_hint %}
-
-                            {% block swag_migration_wizard_page_plugin_information_shopware6_api_hint_plugin %}
-                                <div class="swag-migration-wizard__description swag-migration-profile-shopware6-api-page-information__card">
-                                    <h4 class="information__header">
-                                        {{ $tc('swag-migration.wizard.pages.profileInformation.shopware6Api.plugin.title') }}
-                                    </h4>
-                                    <p>{{ $tc('swag-migration.wizard.pages.profileInformation.shopware6Api.plugin.description') }}</p>
-
-                                    <div class="swag-migration-wizard__link">
-                                        <a class="link"
-                                           target="_blank"
-                                           :href="$tc('swag-migration.wizard.pages.profileInformation.shopware6Api.plugin.link')">
-                                            {{ $tc('swag-migration.wizard.pages.profileInformation.shopware6Api.plugin.linkText') }}
-                                        </a>
-                                    </div>
-                                </div>
-                            {% endblock %}
-
-                            {% block swag_migration_wizard_page_plugin_information_shopware6_api_hint_integration %}
-                                <div class="swag-migration-wizard__description swag-migration-profile-shopware6-api-page-information__card">
-                                    <h4 class="information__header">
-                                        {{ $tc('swag-migration.wizard.pages.profileInformation.shopware6Api.integration.title') }}
-                                    </h4>
-                                    <p>{{ $tc('swag-migration.wizard.pages.profileInformation.shopware6Api.integration.description') }}</p>
-
-                                    <div class="swag-migration-wizard__link">
-                                        <a class="link"
-                                           target="_blank"
-                                           :href="$tc('swag-migration.wizard.pages.profileInformation.shopware6Api.integration.link')">
-                                            {{ $tc('swag-migration.wizard.pages.profileInformation.shopware6Api.integration.linkText') }}
-                                        </a>
-                                    </div>
-                                </div>
-                            {% endblock %}
-                        {% endblock %}
-                    </div>
-                {% endblock %}
-            </div>
-        {% endblock %}
-    </div>
-{% endblock %}
