--- src/Resources/app/administration/src/module/swag-migration/profile/shopware6/api/swag-migration-profile-shopware6-api-page-information/swag-migration-profile-shopware6-api-page-information.scss
@@ -1,31 +1,0 @@
-@import "~scss/variables";
-
-.swag-migration-profile-shopware6-api-page-information {
-    .swag-migration-profile-shopware6-api-page-information__icon {
-        pointer-events: none;
-        display: inline-block;
-        margin-bottom: var(--scale-size-16);
-
-        .sw-settings-item__icon {
-            margin: 0;
-            padding: var(--scale-size-12);
-            border-radius: var(--border-radius-xs);
-            background-color: $color-neutrals-slate-100;
-        }
-    }
-
-    .swag-migration-profile-shopware6-api-page-information__card {
-        border: 1px solid $color-gray-300;
-        border-radius: var(--border-radius-m);
-        padding: var(--scale-size-16);
-        margin-bottom: var(--scale-size-24);
-
-        .information__header {
-            margin-bottom: 0;
-        }
-
-        .swag-migration-wizard__link {
-            margin-top: var(--scale-size-12);
-        }
-    }
-}
