+++ src/Resources/app/administration/src/module/swag-migration/snippet/de-DE.json
@@ -73,34 +73,20 @@
             "local": "Beachte: Die Shopsysteme müssen auf demselben Server liegen. Du benötigst die Zugangsdaten der Datenbank."
           }
         },
-        "profileInformation": {
-          "title": "Migration - Verbindung vorbereiten",
-          "shopware55Api": {
-            "contentTitle": "Installiere den Migration Connector",
-            "contentInformation": "Für die Migration ist die Installation des kostenlosen Plugins „Migration Connector“ erforderlich. Dieses findest Du in Deinem Shopware 5 Backend unter: Plugin Manager > Erweiterungen > Migrationstools. Bitte installiere und aktiviere das Plugin.",
-            "pluginLink": "Mehr Informationen zum Plugin"
-          },
-          "shopware6Api": {
-            "contentTitle": "Anforderungen",
-            "plugin": {
-              "title": "Erweiterung",
-              "description": "Installiere die Migrations-Assistent-Erweiterung im Quellsystem.",
-              "linkText": "Mehr zur Migrations-Assistent-Erweiterung",
-              "link": "https://store.shopware.com/swag257162657297f/migrations-assistent.html"
-            },
-            "integration": {
-              "title": "Integration",
-              "description": "Lege eine Integration im Quellsystem unter Einstellungen > System > Integrationen an. Im nächsten Schritt benötigst Du die dort erzeugte Zugangs-ID und den Sicherheitsschlüssel.",
-              "linkText": "Mehr zu Integrationen",
-              "link": "https://docs.shopware.com/de/shopware-6-de/settings/system/integrationen"
-            }
-          }
-        },
         "credentials": {
           "title": "Migration - Verbindung bearbeiten",
           "emptyState": "Leider wurde kein enstprechendes Formular für dieses Profil gefunden.",
           "shopware55": {
             "api": {
+              "title": "Connect your Shopware 5 store",
+              "steps": {
+                "step1": "Install and activate the plugin <b>Migration Connector</b> from your Shopware 5 backend under Plugin Manager > Extensions > Migration tools.",
+                "step2": "Go to User administration in your Shopware 5 backend.",
+                "step3": "Select the user you want to use for the connection.",
+                "step4": "Activate API access and save.",
+                "step5": "Make sure the user has the role local_admin.",
+                "step6": "Enter the API credentials here to establish the connection."
+              },
               "contentInformation": "Wir benötigen noch einige Daten um eine Verbindung mit Deinem Shopware 5 Shop herstellen zu können. Diese Daten findest Du in der Benutzerverwaltung von Deinem Shopware 5 Backend.",
               "apiKeyLabel": "API-Schlüssel",
               "apiKeyTooltip": "Der API-Schlüssel dient der Sicherheit und Identifizierung.",
@@ -110,7 +96,8 @@
               "apiUserPlaceholder": "Login-Benutzernamen eintragen",
               "endpointLabel": "Shop-Domain",
               "endpointPlaceholder": "www.demoshop.com",
-              "endpointSSLLabel": "Mein Shop verwendet eine sichere Verbindung (SSL)"
+              "endpointSSLLabel": "Mein Shop verwendet eine sichere Verbindung (SSL)",
+              "pluginLinkText": "Mehr Informationen zum Plugin"
             },
             "local": {
               "contentInformation": "Wir benötigen noch Zugangsdaten um eine Verbindung zu Deiner Datenbank von Shopware 5 herstellen zu können.",
@@ -129,6 +116,13 @@
           },
           "shopware6": {
             "api": {
+              "title": "Connect your Shopware 6 store",
+              "steps": {
+                "step1": "Install and activate the <b>Migration Assistant</b> extension on the source system.",
+                "step2": "Create an integration in the source system via Settings > System > Integrations.",
+                "step3": "Generate an access key ID and secret access key.",
+                "step4": "Enter the access keys and shop domain here to establish the connection."
+              },
               "contentInformation": "Du benötigst gültige API-Anmeldeinformationen, um eine Verbindung zwischen Deiner Quell- und Ziel-Instanz von Shopware 6 herzustellen. Diese erhälst Du, wenn Du eine Integration in Deinem Quellsystem erstellst.",
               "apiPasswordLabel": "Sicherheitsschlüssel",
               "apiPasswordTooltip": "Der Sicherheitsschlüssel dient der Sicherheit und Identifizierung.",
@@ -138,7 +132,11 @@
               "apiUserPlaceholder": "Zugangs-ID eintragen",
               "endpointLabel": "Shop-Domain",
               "endpointPlaceholder": "www.demoshop.com",
-              "endpointSSLLabel": "Mein Shop verwendet eine sichere Verbindung (SSL)"
+              "endpointSSLLabel": "Mein Shop verwendet eine sichere Verbindung (SSL)",
+              "pluginLinkText": "Mehr zur Migrations-Assistent-Erweiterung",
+              "pluginLink": "https://store.shopware.com/swag257162657297f/migrations-assistent.html",
+              "integrationLinkText": "Mehr zu Integrationen",
+              "integrationLink": "https://docs.shopware.com/de/shopware-6-de/settings/system/integrationen"
             }
           },
           "statusTitle": "Verbindungscheck",
